pyproject.toml
[project]
name = "agentpro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncpg>=0.31.0",
    "chromadb>=1.5.2",
    "deepagents>=0.4.5",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.2.2",
    "websockets>=16.0",
]
